[minwoo]Fix: requestDto, responseDto 수정

See merge request s10-webmobile1-sub2/S10P12A706!45

+++ backend/saessak/src/main/java/com/ssafy/saessak/board/dto/BoardDetailDto.java
@@ -15,7 +15,7 @@ public class BoardDetailDto {
     Date boardDate;
     String boardContent;
     String boardTemperature;
-    Integer boardSleepTime;
+    Float boardSleepTime;
     String boardPoopStatus;
     Float boardTall;
     Float boardWeight;
